bug: not fixed timeLines ....... :(
