+++ frontend/src/cache.js
@@ -4,6 +4,7 @@ export const initialState = {
     kind: '',
     labels: [],
     comments: '',
+    timeLines: []
   }
 }
