+++ frontend/src/components/InputSite.jsx
@@ -1,3 +1,4 @@
+import axios from 'axios'
 import { useState } from 'react'
 import { useStateMachine } from 'little-state-machine'
 import { Box, Button, Chip, FormControl, Grid} from '@material-ui/core'
@@ -14,6 +15,7 @@ const InputSite = (props) => {
   const [site, setSite] = useState(initialCache.site)
   const [kind, setKind] = useState(initialCache.kind)
   const [labels, setLabels] = useState(initialCache.labels)
+  const [timeLines, setTimeLines] = useState(initialCache.timeLines)
   //
   const labelList = [
     'AWS', 'Azure', 'Linux', 'Git', 'Docker', 'Python', 'Javascript', 'TypeScript', 'React', 'Other', 'Column'
@@ -37,8 +39,23 @@ const InputSite = (props) => {
     actions.updateContent({labels})
   }
   //
-  const handleSubmit = () => {
-    actions.updateContent({site,kind})
+  const handleSubmit = async(event) => {
+    event.preventDefault()
+    const getNum = 5
+    const apiPath = `https://ftcg0rr8h3.execute-api.ap-northeast-1.amazonaws.com/api/history/${getNum}`
+    const { data } = await axios.get(apiPath)
+    const results = data.body
+    for (let i = 0; i < results.length; i++){
+     const iobj = {
+        title: results[i].date,
+        cardTitle: results[i].site,
+        cardSubtitle: results[i].labels,
+        cartDetailedTitle: results[i].comments
+      }
+      setTimeLines([...timeLines, iobj])
+      console.log(iobj)
+    }
+    actions.updateContent({site,kind,timeLines})
     props.handleNext()
   }
   return (
